# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de BarberMusic&Spa",
        "description": "Documentación para la API utilizada en el sistema de BarberMusic&Spa. Aquí puedes probar los endpoints de autenticación.",
        "contact": {
            "email": "soporte@barbermusic.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/Admin_personal/personal": {
            "get": {
                "tags": [
                    "Personal"
                ],
                "summary": "Obtener todo el personal",
                "operationId": "72ae6b237f991299b826e067e362da9c",
                "responses": {
                    "200": {
                        "description": "Lista del personal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Personal"
                ],
                "summary": "Crear un nuevo registro de personal",
                "operationId": "5c03022964284c411f6de4188dc70761",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "usuario_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "sucursal_asignada_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo_personal": {
                                        "type": "string",
                                        "example": "BARBERO"
                                    },
                                    "numero_empleado": {
                                        "type": "string",
                                        "example": "EMP-001"
                                    },
                                    "fecha_contratacion": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-15"
                                    },
                                    "activo_en_empresa": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Personal creado exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Personal creado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/Admin_personal/personal/{id}": {
            "get": {
                "tags": [
                    "Personal"
                ],
                "summary": "Obtener un registro de personal por ID",
                "operationId": "7043f200ff396709bf72dac922086167",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del personal"
                    },
                    "404": {
                        "description": "Personal no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Personal"
                ],
                "summary": "Actualizar un registro de personal",
                "operationId": "39a6d4039d568f042d3bd494043131ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sucursal_asignada_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "tipo_personal": {
                                        "type": "string",
                                        "example": "ESTILISTA"
                                    },
                                    "activo_en_empresa": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Personal actualizado"
                    },
                    "404": {
                        "description": "Personal no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Personal"
                ],
                "summary": "Eliminar un registro de personal (Soft Delete)",
                "operationId": "8b47946d4bcb1eaf71b24aa8a58ede6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Personal eliminado"
                    },
                    "404": {
                        "description": "Personal no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce el token de acceso (JWT o Sanctum) sin la palabra 'Bearer'.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}